requirements.txt
rich>=13.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
markdown>=3.4.0
pypdf>=3.0.0
pdfplumber>=0.10.0